+++ dev/examples.el
@@ -625,16 +625,33 @@ value rather than consuming a list to produce a single value."
   "Reductions of one or more lists to a boolean value."
 
   (defexamples -any?
-    (-any? 'even? '(1 2 3)) => t
-    (-any? 'even? '(1 3 5)) => nil
-    (-any? 'null '(1 3 5)) => nil
-    (-any? 'null '(1 3 ())) => t
-    (--any? (= 0 (% it 2)) '(1 2 3)) => t)
+    (-any? #'numberp '(nil 0 t)) => t
+    (-any? #'numberp '(nil t t)) => nil
+    (-any? #'null '(1 3 5)) => nil
+    (-any? #'null '(1 3 ())) => t
+    (-any? #'identity '()) => nil
+    (-any? #'identity '(0)) => t
+    (-any? #'identity '(nil)) => nil
+    (--any? (= 0 (% it 2)) '(1 2 3)) => t
+    (--any? (= it it-index) '()) => nil
+    (--any? (= it it-index) '(0)) => t
+    (--any? (= it it-index) '(1)) => nil
+    (--any? (= it it-index) '(1 1)) => t
+    (--any? (= it it-index) '(1 2)) => nil)
 
   (defexamples -all?
-    (-all? 'even? '(1 2 3)) => nil
-    (-all? 'even? '(2 4 6)) => t
-    (--all? (= 0 (% it 2)) '(2 4 6)) => t)
+    (-all? #'numberp '(1 2 3)) => t
+    (-all? #'numberp '(2 t 6)) => nil
+    (--all? (= 0 (% it 2)) '(2 4 6)) => t
+    (-all? #'identity '()) => t
+    (-all? #'identity '(0)) => t
+    (-all? #'identity '(0 1)) => t
+    (-all? #'identity '(nil)) => nil
+    (--all? (= it it-index) '()) => t
+    (--all? (= it it-index) '(0)) => t
+    (--all? (= it it-index) '(1)) => nil
+    (--all? (= it it-index) '(1 1)) => nil
+    (--all? (= it it-index) '(0 1)) => t)
 
   (defexamples -none?
     (-none? 'even? '(1 2 3)) => nil
@@ -1086,7 +1103,9 @@ related predicates."
     (-first #'identity '()) => nil)
 
   (defexamples -some
-    (-some (lambda (s) (string-match-p "x" s)) '("foo" "axe" "xor")) => 1
+    (-some #'stringp '(1 "2" 3)) => t
+    (--some (string-match-p "x" it) '("foo" "axe" "xor")) => 1
+    (--some (= it-index 3) '(0 1 2)) => nil
     (-some (lambda (s) (string-match-p "x" s)) '("foo" "bar" "baz")) => nil
     (--some (member 'foo it) '((foo bar) (baz))) => '(foo bar)
     (--some (plist-get it :bar) '((:foo 1 :bar 2) (:baz 3))) => 2
@@ -1102,6 +1121,32 @@ related predicates."
     (--some it '(1 2 3)) => 1
     (--some it '(1)) => 1
     (--some it '()) => nil)
+
+  (defexamples -every
+    (-every #'numberp '(1 2 3)) => t
+    (--every (string-match-p "x" it) '("axe" "xor")) => 0
+    (--every (= it it-index) '(0 1 3)) => nil
+    (-every #'ignore '()) => t
+    (-every #'ignore '(0)) => nil
+    (-every #'ignore '(0 1)) => nil
+    (--every nil '()) => t
+    (--every nil '(0)) => nil
+    (--every nil '(0 1)) => nil
+    (-every #'identity '()) => t
+    (-every #'identity '(0)) => 0
+    (-every #'identity '(0 1)) => 1
+    (--every it '()) => t
+    (--every it '(1)) => 1
+    (--every it '(1 2)) => 2
+    (--every it-index '()) => t
+    (--every it-index '(1)) => 0
+    (--every it-index '(1 2)) => 1
+    (let ((r 'r)) (-every (lambda (x) (setq r x)) '()) r) => 'r
+    (let ((r 'r)) (-every (lambda (x) (setq r x)) '(nil 1)) r) => nil
+    (let (r) (-every (lambda (x) (setq r x)) '(0 1)) r) => 1
+    (let (i) (--every (ignore (setq i it-index)) '()) i) => nil
+    (let (i) (--every (ignore (setq i it-index)) '(a)) i) => 0
+    (let (i) (--every (ignore (setq i it-index)) '(a b)) i) => 0)
 
   (defexamples -last
     (-last 'even? '(1 2 3 4 5 6 3 3 3)) => 6